Add a flag to specify an alternate module root. (#51)
Although it is common practice, not all Go repositories have the go.mod file at
the repo root. Add a --moduleRoot flag that allows the caller to specify a path
relative to the repository root, that should be used as the module root for the
purpose of indexing.

To handle vendored dependencies, we now use go list -mod=readonly.
To preserve the existing assumption that dependencies are fetched, do an
explicit go mod download before listing the module.