indexer: prepare to index documentSymbols (do not index duplicate definitions)
This prepares the indexer to index `textDocument/documentSymbols`. The problem
here is that indexing symbols involves emitting ranges for the symbol definitions,
and so does our existing indexing for `indexDefinition`. LSIF ranges cannot overlap
or be duplicative, so we must effectively make symbol and definition indexing work
together. The idea is that symbol indexing will emit ranges for the definitions it
encounters, and traditional definition indexing will merely reuse those or emit a
new range if needed.

At present, this causes all tests to fail due to the way we pre-emptively mark
a range for generation during definition indexing:

```
--- FAIL: TestIndexer (0.63s)
    --- FAIL: TestIndexer/check_Parallel_function_hover_text (0.00s)
        indexer_test.go:35: could not find target range
    --- FAIL: TestIndexer/check_external_package_hover_text (0.00s)
        indexer_test.go:62: could not find target range
    --- FAIL: TestIndexer/check_errs_definition (0.00s)
        indexer_test.go:94: incorrect definition count. want=1 have=0
    --- FAIL: TestIndexer/check_wg_references (0.00s)
        indexer_test.go:108: incorrect reference count. want=4 have=3
    --- FAIL: TestIndexer/check_NestedB_monikers (0.00s)
        indexer_test.go:122: could not find target range
    --- FAIL: TestIndexer/check_typeswitch (0.00s)
        indexer_test.go:143: could not find target range
    --- FAIL: TestIndexer/check_typealias (0.00s)
        indexer_test.go:209: could not find target range
    --- FAIL: TestIndexer/check_typealias_reference (0.00s)
        indexer_test.go:255: incorrection definition count. want=1 have=0
    --- FAIL: TestIndexer/check_typealias_anonymous_struct (0.00s)
        indexer_test.go:296: could not find target range
FAIL
FAIL	github.com/sourcegraph/lsif-go/internal/indexer	17.393s
FAIL
```

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>