indexer: remove pre-emptive range marking for definitions
Prior to this change, we would pre-emptively "mark" a range for a definition
as being actively produced. This appears to have been because `indexDefinition`
would _always_ emit a new range, and this it was necessary to ensure that two
parallel definition indexing processes would not emit the same range.

Since my prior commit, however, `indexDefinition` only emits a new range if one
does not exist. So pre-emptive range marking is not required anymore.

This causes all tests to pass.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>