API docs: emit directory-structured pages
Prior to this change, we would emit documentation pages with a completely flat structure
in the workspace like:

* `internal-lib-protocol`
* `internal-lib-util`
* `pkg-mux`
* `pkg-router`

Which meant the index page for a workspace would just be a giant list of all packages in
the workspace. Instead, we now emit documentation pages matching the package import path
directory structure, e.g.:

* / (index)
    * internal (index)
        * lib (index)
            * protocol
            * util
    * pkg (index)
        * mux
        * router

Which makes it possible to navigate the workspace's Go package documentation as a tree instead
of as one giant flat list.

Helps sourcegraph/sourcegraph#21922

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>