API docs: emit many more tags (#175)
This implements sourcegraph/sourcegraph#22327 by causing us to tag documentation for e.g. symbols with more information about the type of symbol being described. e.g. for a const foo = 5 we now emit tags [constant number private] to signal that it is a constant, a number, and a private symbol.

These tags are useful for differentiating things in the UI with e.g. icons, as well as for filtering and searching documentation to just the kinds of things you are interested in.

Helps https://github.com/sourcegraph/sourcegraph/issues/22326

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>