gomod: special case listing the stdlib
Before this commit lsif-go fails on the standard library for go. When
running the "go list all" command fails to run with an error message
like:

  cannot find package "." in:
    path/to/a/vendor/dir
  ...

This might be an issue in upstream go list. However, you normally do
need special handling of the go stdlib tool. In our case we can just
skip listing dependencies since it is the standard library.

Test Plan: We eyeballed the lsif dump after running on stdlib, and it
looked legitimate. However, we don't know much about lsif.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>